+++ manifest.config.ts
@@ -36,6 +36,7 @@ export default defineManifest(async (env) => ({
         "src/character-tools.js",
         "src/options-effect.js",
         "src/incrementor-tools.js",
+        "src/target-tools.js",
         "src/leaderboard-tools.js",
         "src/dom-tools.js",
         "src/mode-menu.js",